+++ include/boost/beast/src.hpp
@@ -52,6 +52,7 @@ the program, with the macro BOOST_BEAST_SEPARATE_COMPILATION defined.
 #include <boost/beast/http/impl/verb.ipp>
 
 #include <boost/beast/websocket/detail/hybi13.ipp>
+#include <boost/beast/websocket/detail/mask.ipp>
 #include <boost/beast/websocket/detail/pmd_extension.ipp>
 #include <boost/beast/websocket/detail/prng.ipp>
 #include <boost/beast/websocket/detail/service.ipp>