+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 260:
 
 * More split compilation in rfc7230.hpp
 * Qualify calls to `beast::iequals` in basic_parser.ipp
+* More split compilation in websocket/detail/mask.hpp
 
 --------------------------------------------------------------------------------